+++ src/Fumo.Commands/BotCommand.cs
@@ -2,8 +2,8 @@
 using Fumo.Database.Extensions;
 using Fumo.Shared.Enums;
 using Fumo.Shared.Exceptions;
-using Fumo.Shared.Interfaces;
 using Fumo.Shared.Models;
+using Fumo.Shared.Repositories;
 using Fumo.Shared.ThirdParty.Pajbot1;
 
 namespace Fumo.Commands;
