+++ src/Fumo.Commands/SevenTV/SevenTVSearchCommand.cs
@@ -1,11 +1,11 @@
 ﻿using Fumo.Shared.Exceptions;
-using Fumo.Shared.Interfaces;
 using Fumo.Shared.Models;
 using Fumo.Shared.Regexes;
 using Fumo.Shared.ThirdParty.Emotes.SevenTV;
 using Fumo.Shared.ThirdParty.Exceptions;
 using System.Text;
 using Fumo.Shared.ThirdParty.Emotes.SevenTV.Models;
+using Fumo.Shared.Repositories;
 
 namespace Fumo.Commands.SevenTV;
