+++ src/Fumo.Commands/SevenTV/SevenTVYoinkCommand.cs
@@ -1,13 +1,13 @@
 ﻿using Fumo.Database;
 using Fumo.Database.Extensions;
 using Fumo.Shared.Exceptions;
-using Fumo.Shared.Interfaces;
 using Fumo.Shared.Models;
 using Fumo.Shared.Utils;
 using Fumo.Shared.ThirdParty.Emotes.SevenTV;
 using System.Collections.Immutable;
 using Fumo.Shared.ThirdParty.Emotes.SevenTV.Enums;
 using Fumo.Shared.ThirdParty.Emotes.SevenTV.Models;
+using Fumo.Shared.Repositories;
 
 namespace Fumo.Commands.SevenTV;
