+++ src/Fumo.Shared/Mediator/EventsubChatMessageCommand.cs
@@ -3,8 +3,8 @@ using Fumo.Database;
 using Fumo.Database.DTO;
 using Fumo.Database.Extensions;
 using Fumo.Shared.Eventsub;
-using Fumo.Shared.Interfaces;
 using Fumo.Shared.Models;
+using Fumo.Shared.Repositories;
 using MediatR;
 using MiniTwitch.Helix.Responses;
 using MiniTwitch.Irc;
